[sonic_yang_ext.py]: Rename _sonic_yang_ext.py as sonic_yang_ext.py.
