+++ src/sonic-yang-mgmt/setup.py
@@ -90,7 +90,7 @@ setup(
     include_package_data=True,
     keywords='sonic_yang_mgmt',
     name='sonic_yang_mgmt',
-    py_modules=['sonic_yang', '_sonic_yang_ext'],
+    py_modules=['sonic_yang', 'sonic_yang_ext'],
     packages=find_packages(),
     setup_requires=setup_requirements,
     version='1.0',
